LayoutTests/inspector/debugger/script-formatter-breakpoints-expected.txt

Tests the script formatting is working fine with breakpoints.

Debugger was enabled.

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in raw

    script-formatter-breakpoints.html:11
    var f=0;
Script execution resumed.
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

    script-formatter-breakpoints.html:18
    var f = 0;
Breakpoint sidebar pane while paused in raw

    script-formatter-breakpoints.html:11
    var f=0;
Script execution resumed.
Debugger was disabled.
